Fixing Sileo/Installer Depictions
Manually fixing an error with the packages that removed their Sileo/Installer depictions from the packages file. Will have to correct this in the deb files later so that this no longer has to be correctted manually but this should work for the time being. Also corrected a typo in the Stoned Penguin Sileo/Installer depiction.